Fixed bug in insertAataToMySql

+++ arduino/smartHouse/smartHouse.ino
@@ -181,8 +181,8 @@ void loop()
     if (gNumberOfTimeDomains <= passedTimeDomains)
     {            
        insertDataToMySql(
-       (unsigned long)(tempForSql * 100 /(gNumberOfTimeDomains+1)),
-       (unsigned long)(humForSql  * 100 /(gNumberOfTimeDomains+1)),
+       (unsigned int)(tempForSql * 100 /(gNumberOfTimeDomains+1)),
+       (unsigned int)(humForSql  * 100 /(gNumberOfTimeDomains+1)),
        (unsigned long)(presForSql* 100 /(gNumberOfTimeDomains+1)),
        (unsigned long)(altForSql * 100 /(gNumberOfTimeDomains+1)));
